+++ NEWS.md
@@ -1,5 +1,10 @@
 # dplyr 0.7.1.9000
 
+* Fix protection error that occurred when creating a character column
+  using grouped `mutate()` (#2971).
+
+* `distinct()` now throws an error when used on unknown columns (#2867, @foo-bar-baz-qux).
+
 # dplyr 0.7.1
 
 * Use new versions of bindrcpp and glue to avoid protection problems.